LLMReceiptPrintAPI/requirements.txt
Flask~=3.1.3
ollama~=0.6.1
pyusb~=1.3.1